--- mentionserver/src/main/java/com.mention/dao/FollowDao.java
@@ -1,7 +1,0 @@
-package com.mention.dao;
-
-import com.mention.model.Follow;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FollowDao extends JpaRepository<Follow, Long> {
-}
