+++ mentionserver/src/main/java/com.mention/repository/FollowRepository.java
@@ -1,0 +1,7 @@
+package com.mention.repository;
+
+import com.mention.model.Follow;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FollowRepository extends JpaRepository<Follow, Long> {
+}
