+++ mentionserver/src/main/java/com.mention/service/FollowServiceImpl.java
@@ -1,6 +1,6 @@
 package com.mention.service;
 
-import com.mention.dao.FollowDao;
+import com.mention.repository.FollowRepository;
 import com.mention.model.Follow;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,28 +10,28 @@ import java.util.Optional;
 @Service
 public class FollowServiceImpl implements FollowService {
 
-  private FollowDao followDao;
+  private FollowRepository followRepository;
 
-  public FollowServiceImpl(FollowDao followDao) {
-    this.followDao = followDao;
+  public FollowServiceImpl(FollowRepository followRepository) {
+    this.followRepository = followRepository;
   }
 
   @Override
   @Transactional
   public void addFollow(Follow follow) {
-    followDao.save(follow);
+    followRepository.save(follow);
   }
 
 
   @Override
   @Transactional
   public void deleteFollow(Long id) {
-    followDao.deleteById(id);
+    followRepository.deleteById(id);
 
   }
 
   @Override
   public Optional<Follow> getFollow(Long id) {
-    return followDao.findById(id);
+    return followRepository.findById(id);
   }
 }
